Fix a symbol error, re-run minisvd2nim
The .svd file had a comma (,) in a name element
which resulted in an invalid Nim symbol.
I changed that element to be more like the register
that immediately preceded it.

I re-ran minisvd2nim on the .svd since the tool has had
meaningful updates.

I will also change the minisvd2nim tool to prevent problems
with commas in elements that are mapped straight to Nim symbols.d